phabricator: Deploy user should own everything under old rev directories
Transfer ownership of old rev directories to the deploy user following
`scap deploy`'s promote stage. This allows scap to perform its clean up
(removing old rev trees) during the finalize stage.

In addition, refactored all phabricator deploy scripts to be files
instead of ERB templates and moved variables to a central ERB template
per policy.

Bug: T313953

+++ modules/phabricator/files/phab_deploy_promote.sh
@@ -1,4 +1,31 @@
 #!/bin/sh
+#
+# Runs as a scap deploy check following the promote stage.
+#
+
+. /etc/phabricator/script-vars
+
+if [ -z "$SCAP_REVS_DIR" ]; then
+  echo '$SCAP_REVS_DIR is not defined.'
+  echo 'Note: This script is only intended to run as a scap deploy check'
+  exit 1
+fi
 
 systemctl stop phd
-puppet agent --disable
+disable-puppet 'phabricator deployment'
+
+# Transfer ownership to the deploy user for all files under the old rev
+# directories to allow scap to clean up during the finalize stage.
+#
+# -type d: directories
+# -mindepth 1: avoid including the starting dir
+# -maxdepth 1: only top-level directories
+# -not -path: exclude this rev path, scap's "current" rev, and scap's "done" rev
+
+find "$SCAP_REVS_DIR" \
+  -type d -maxdepth 1 -mindepth 1 \
+  -not -path "$SCAP_REV_PATH" \
+  -not -path "$SCAP_CURRENT_REV_DIR" \
+  -not -path "$SCAP_DONE_REV_DIR" \
+  -exec sudo chown -R "$PHAB_DEPLOY_USER" '{}' \;
+  -exec sudo chmod -R u+w '{}' \;
